backend_fastapi/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
asyncpg==0.30.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.11.7
python-multipart==0.0.20
